We want more (f64)
This replaces a bunch of use of f32 (in parsing and in coordinate types)
with f64.

There are two tests that are failing which I don't want to touch.

This doesn't have a dramatic impact on crater, but it does show
improvements for fvar & HVAR for a number of fonts.

+++ fea-rs/src/compile/validate.rs
@@ -249,7 +249,7 @@ impl<'a, V: VariationInfo> ValidationCtx<'a, V> {
             self.error(condition.tag().range(), "unknown axis");
             return;
         };
-        if (condition.min_value().parse_signed() as f64) < axis.min.into_inner().0 as f64 {
+        if (condition.min_value().parse_signed() as f64) < axis.min.into_inner().0 {
             self.error(
                 condition.min_value().range(),
                 format!(
@@ -258,7 +258,7 @@ impl<'a, V: VariationInfo> ValidationCtx<'a, V> {
                 ),
             );
         }
-        if (condition.max_value().parse_signed() as f64) > axis.max.into_inner().0 as f64 {
+        if (condition.max_value().parse_signed() as f64) > axis.max.into_inner().0 {
             self.error(
                 condition.max_value().range(),
                 format!(
